Add experimental TDX attester/verifier crate (#35)
The az-snp-vtpm crate has been generalized into an az-cvm-vtpm
workspace and extended to support attestation workflows on Azure TDX
CVMs, which are in limited preview at the moment. The attestation
workflow on TDX CVMs has a lot in common with the equivalent process
on SEV-SNP CVMs, so in the future we'll probably want to consolidate
both into a common az-cvm-vtpm crate.

Until TDX CVMs on are generally available on Azure, we'll want to
keep both crates seperate, as the offering is still subject to change.

Signed-off-by: Magnus Kulke <magnuskulke@microsoft.com>