deps: replace rsa and sha2 crates with openssl
fixes #51
fixes #46

The rsa crate is being reported in audit jobs due to a timing-related
security issue. The project is not committed to address this in the near
time, since the solution involves switching to a big-num dependency with
worse security characteristics.

We can leave it up to the consumer of an attester-only crate to convert
the pubkey into a library struct.

Signed-off-by: Magnus Kulke <magnuskulke@microsoft.com>